+++ src/redturtle/prenotazioni/browser/week.py
@@ -9,9 +9,15 @@ from redturtle.prenotazioni.browser.base import BaseView
 from redturtle.prenotazioni.browser.interfaces import IDontFollowMe
 from redturtle.prenotazioni.utilities.urls import urlify
 from six.moves import range
+from email.utils import formatdate
 from zope.deprecation import deprecate
 from zope.interface import implementer
 from zope.schema.vocabulary import getVocabularyRegistry
+
+import time
+
+TIPOLOGIA_PRENOTAZIONE_NAME = "TipologiaPrenotazione"
+TIPOLOGIA_PRENOTAZIONE_NAME_COOKIE = "TipologiaPrenotazione_cookie"
 
 
 @implementer(IDontFollowMe)
@@ -145,6 +151,12 @@ class View(BaseView):
         """The link to the previous week"""
         qs = {"data": self.prev_week}
         qs.update(self.prenotazioni.remembered_params)
+
+        prenotation_type = self.get_prenotation_type()
+
+        if prenotation_type:
+            qs[TIPOLOGIA_PRENOTAZIONE_NAME] = prenotation_type
+
         return urlify(self.request.getURL(), params=qs)
 
     @property
@@ -153,6 +165,12 @@ class View(BaseView):
         """The link to the next week"""
         qs = {"data": self.next_week}
         qs.update(self.prenotazioni.remembered_params)
+
+        prenotation_type = self.get_prenotation_type()
+
+        if prenotation_type:
+            qs[TIPOLOGIA_PRENOTAZIONE_NAME] = prenotation_type
+
         return urlify(self.request.getURL(), params=qs)
 
     @property
@@ -272,8 +290,31 @@ class View(BaseView):
         )
         return message
 
+    def get_prenotation_type(self):
+        """Returns the prenotation type passed by url"""
+        return self.request.get(TIPOLOGIA_PRENOTAZIONE_NAME, "")
+
+    def set_cookies(self):
+        self.set_prenotation_type()
+
+    def set_prenotation_type(self):
+        prenotation_type = self.get_prenotation_type()
+        cookie_value = self.request.cookies.get(TIPOLOGIA_PRENOTAZIONE_NAME_COOKIE, "")
+
+        if prenotation_type and not cookie_value:
+            # set expiring cookies
+            expiration_seconds = time.time() + (30 * 60)
+            expires = formatdate(expiration_seconds, usegmt=True)
+            self.request.response.setCookie(
+                TIPOLOGIA_PRENOTAZIONE_NAME_COOKIE,
+                prenotation_type.strip().encode("utf-8"),
+                expires=expires,
+                path="/",
+            )
+
     def __call__(self):
         """Hide the portlets before serving the template"""
         # self.request.set('disable_plone.leftcolumn', 1)
         # self.request.set('disable_plone.rightcolumn', 1)
+        self.set_cookies()
         return super(View, self).__call__()