+++ src/redturtle/prenotazioni/upgrades.py
@@ -7,6 +7,11 @@ from plone.app.upgrade.utils import loadMigrationProfile
 from plone.app.workflow.remap import remap_workflow
 from plone.contentrules.engine.interfaces import IRuleStorage
 from zope.component import queryUtility
+from plone.app.contentrules.conditions.wfstate import WorkflowStateCondition
+from plone.app.contentrules.conditions.wftransition import WorkflowTransitionCondition
+from plone.app.contentrules.actions.workflow import WorkflowAction
+from plone.app.contentrules.conditions.portaltype import PortalTypeCondition
+from plone import api
 
 import logging
 
@@ -163,3 +168,23 @@ def to_1400(context):
 
 def to_1401(context):
     update_types(context)
+
+    rule_storage = queryUtility(IRuleStorage)
+    for rule in rule_storage.items():
+        rule = rule[1]
+
+        workflow_action_conditions = filter(
+            lambda item: isinstance(item, WorkflowAction), rule.actions
+        )
+        portal_type_conditions = filter(
+            lambda item: isinstance(item, PortalTypeCondition), rule.conditions
+        )
+
+        if [
+            condition
+            for condition in portal_type_conditions
+            if "Prenotazione" in getattr(condition, "check_types", [])
+        ]:
+            for workflow_action in workflow_action_conditions:
+                if workflow_action.transition == "publish":
+                    workflow_action.transition = "confirm"